requirements.txt
ddtrace
openai
requests==2.32.2
urllib3==1.26.15
notebook==7.1.3
python-dotenv==1.0.1
black==24.4.2
